BroadcastMoinitorCall.java - before using access to Modulexxx.self(), not null check is asserted
BaseModule.java
haveStops() check list emptiness and in case of stops print a log

ModuleCamera.java
Added a log print

ModuleMic.java
When a process in backlist is started, addStop/removeStop is called.

CameraSnapshot.java,Instrument.java,ModuleMic.java
For all the Modules which kills the mediaserver, a 2secs sleep has been added on order to keep locked the semaphoreMediaserver.
In this way the moduleMic knows when mediaserver has been killed by RCS