UninstallAction.java:  added comment
Root.java:
Uninstall procedure enachement, removed use of executeRootAndForgetScript for the uninstall script
do to some unexpected result when Persistent and not rebooted. In that case the script wasn't executed.
In the uninstallation scrip the app is force-stopped to avoid problem in /systm/app/Stk* removal for "File busy error".